fix: Streamlit fixes (#14)
Fixes #7
Moved most code to streamlit_deephaven.py and removed DEV_MODE since running test.py directly is no more difficult.

Fixes #8
Fixed through moving server start code to a cache resource to ensure it is only ran once.

Fixes #9
Added _remove_widgets(), which is called every time the server reruns.

Fixes #10
Added a random nonce. I'm not sure if it really matters but I used a completely random one rather than the id of the widget as the id of the widget is the memory address in CPython.